Get rid of StdLib rests Removed ./Debug/ from repo Asm delay using timer (just for fun) Added si4705 properties support Digital Audio clock generation ousting TIM1

+++ include/delay.h
@@ -9,7 +9,7 @@
 #define DELAY_H_
 
 void delay_init();
-void delay_ms(uint32_t ms);
-void delay_us(uint32_t us);
+void delay_ms(uint16_t ms);
+void delay_us(uint16_t us);
 
 #endif /* DELAY_H_ */
